feat(#5) : CoupleMap 컴포넌트 생성
- naver map init
- 컴포넌트명 변경
